Add KL divergence between Haar and VQC, and scale expressibility
This commits adds a plot for the KL divergence between the Haar random
probability distribution and the VQC's fidelity probability
distribution.

Previously, the expressibility diagram contained true probability
distribution histograms, in the sense that the histogram's area (i.e. the
derivative) is one. This commit adopts this to be closer to the
histograms by Sim et al., calculating the absolute histograms and
dividing by the number of samples.